+++ Editor/PackageConfig.cs
@@ -3,6 +3,6 @@ namespace Hananoki.GitHubDownload {
 	public static class PackageInfo {
 		public const string name = "GitHubDownload";
 		public const string editorPrefName = "Hananoki.GitHubDownload";
-		public const string version = "0.4.1-preview";
+		public const string version = "0.5.0-preview";
 	}
 }